refactor(tele): rework structure
