+++ core/src/cn/harryh/arkpets/telemetry/PrivacyField.java
@@ -1,4 +1,4 @@
-package cn.harryh.arkpets;
+package cn.harryh.arkpets.telemetry;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -6,7 +6,7 @@ import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
 
-/** Marks an ArkConfig field that must not be collected into telemetry snapshots.
+/** Marks an config field that must not be collected into telemetry snapshots.
  */
 @Retention(RetentionPolicy.RUNTIME)
 @Target(ElementType.FIELD)
